+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ latest
 * Include py.typed file in package data to support type checking
 * Remove upper bounds on dependencies. This allows usage of Grimp 2.0, which should significantly speed up checking of
   layers contracts.
+* Add --verbose flag to lint-imports command.
 
 1.3.0 (2022-08-22)
 ------------------