+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-latest
-------
+1.4.0 (2022-10-04)
+------------------
 
 * Include py.typed file in package data to support type checking
 * Remove upper bounds on dependencies. This allows usage of Grimp 2.0, which should significantly speed up checking of